% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_2line.R
\name{f_2line}
\alias{f_2line}
\title{Drawing line charts of two categories or single category versus cross-sectional benchmark}
\usage{
f_2line(df, cids, cats, bm = "GLB", dax = FALSE, title = NULL,
  selout = NULL, leg1 = NULL, leg2 = NULL, size = 1, showzero = TRUE,
  devnew = FALSE)
}
\arguments{
\item{df}{zoo dataframe, containing all series of interest}

\item{cids}{character (vector), contains one or several cross-sections}

\item{cats}{character (vector), contains one or two categories}

\item{bm}{character, cross-sectional benchmark, default is "GLB", i.e. average is created (also for 2-categories case)}

\item{dax}{logical, TRUE means that series have two different axes}

\item{title}{character string, optional chart title, default is single cross section or category acronym}

\item{selout}{character (vector), default is NULL, contain the cross-sections for which we want to display the output \cr
i.e. can be diffrent from cids}

\item{leg1}{character string, optional legend for line 1 (blue), default is panel member acronym}

\item{leg2}{character string, optional legend for line 2 (dotted black), default is panel member acronym}

\item{size}{numeric, scales size of title and legend relative to all other graphical elements}

\item{showzero}{boolean, showing thick zero line, default is TRUE}

\item{devnew}{boolean, whether to open new graphic device for each graph (defalut=\code{FALSE} for RStudio use)}
}
\value{

}
\description{
The function \code{f_2line} draws line charts of two categories or single category versus cross-sectional benchmark for cross-sections of a panel
}
\details{

}
\examples{
require(zoo)
dfm <- zoo(matrix(rnorm(288, -5, 5), 48, 6), seq(as.Date("2012-02-01"), length=48, by="1 month") - 1)
names(dfm) <- as.vector(outer(c("EUR", "USD", "JPY"), c("V1", "V2"), paste, sep="_"))
f_2line(dfm, c("EUR", "USD", "JPY"), cats = c("V1", "V2"), dax=TRUE)
f_2line(dfm, c("EUR", "USD", "JPY"), cats = "V1", bm = c("EUR", "JPY"), dax=T)
f_2line(dfm, c("EUR", "USD", "JPY"), cats = "V1", bm = c("EUR"), dax=T)
f_2line(dfm, c("EUR", "USD", "JPY"), cats = c("V1", "V2"), bm = "JPY")
f_2line(dfm, c("EUR", "USD", "JPY"), cats = c("V1", "V2"), bm = "GLB", selout = c("EUR", "GLB"))
f_2line(dfm, c("EUR", "USD", "JPY"), cats = c("V1", "V2"), bm = "GLB", selout = c("GLB"))

}

